requirements.txt
pytest==7.1.2
requests==2.26.0
jsonschema==3.2.0
cerberus==1.3.4
selenium==4.2.0
allure-pytest==2.9.45